use stable blob SDK
