+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_blob_client.py
@@ -52,7 +52,7 @@ from ._quick_query_helper import BlobQueryReader
 from ._upload_helpers import (
     upload_block_blob,
     upload_append_blob,
-    upload_page_blob, _any_conditions)
+    upload_page_blob)
 from ._models import BlobType, BlobBlock, BlobProperties, BlobQueryError
 from ._download import StorageStreamDownloader
 from ._lease import BlobLeaseClient
@@ -406,146 +406,6 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             raise ValueError("Unsupported BlobType: {}".format(blob_type))
         return kwargs
 
-    def _upload_blob_from_url_options(self, source_url, **kwargs):
-        # type: (...) -> Dict[str, Any]
-        tier = kwargs.pop('standard_blob_tier', None)
-        overwrite = kwargs.pop('overwrite', False)
-        content_settings = kwargs.pop('content_settings', None)
-        if content_settings:
-            kwargs['blob_http_headers'] = BlobHTTPHeaders(
-                blob_cache_control=content_settings.cache_control,
-                blob_content_type=content_settings.content_type,
-                blob_content_md5=None,
-                blob_content_encoding=content_settings.content_encoding,
-                blob_content_language=content_settings.content_language,
-                blob_content_disposition=content_settings.content_disposition
-            )
-        cpk = kwargs.pop('cpk', None)
-        cpk_info = None
-        if cpk:
-            if self.scheme.lower() != 'https':
-                raise ValueError("Customer provided encryption key must be used over HTTPS.")
-            cpk_info = CpkInfo(encryption_key=cpk.key_value, encryption_key_sha256=cpk.key_hash,
-                               encryption_algorithm=cpk.algorithm)
-
-        options = {
-            'content_length': 0,
-            'copy_source_blob_properties': kwargs.pop('include_source_blob_properties', True),
-            'source_content_md5': kwargs.pop('source_content_md5', None),
-            'copy_source': source_url,
-            'modified_access_conditions': get_modify_conditions(kwargs),
-            'blob_tags_string': serialize_blob_tags_header(kwargs.pop('tags', None)),
-            'cls': return_response_headers,
-            'lease_access_conditions': get_access_conditions(kwargs.pop('destination_lease', None)),
-            'tier': tier.value if tier else None,
-            'source_modified_access_conditions': get_source_conditions(kwargs),
-            'cpk_info': cpk_info,
-            'cpk_scope_info': get_cpk_scope_info(kwargs)
-        }
-        options.update(kwargs)
-        if not overwrite and not _any_conditions(**options): # pylint: disable=protected-access
-            options['modified_access_conditions'].if_none_match = '*'
-        return options
-
-    @distributed_trace
-    def upload_blob_from_url(self, source_url, **kwargs):
-        # type: (str, Any) -> Dict[str, Any]
-        """
-        Creates a new Block Blob where the content of the blob is read from a given URL.
-        The content of an existing blob is overwritten with the new blob.
-
-        :param str source_url:
-            A URL of up to 2 KB in length that specifies a file or blob.
-            The value should be URL-encoded as it would appear in a request URI.
-            If the source is in another account, the source must either be public
-            or must be authenticated via a shared access signature. If the source
-            is public, no authentication is required.
-            Examples:
-            https://myaccount.blob.core.windows.net/mycontainer/myblob
-
-            https://myaccount.blob.core.windows.net/mycontainer/myblob?snapshot=<DateTime>
-
-            https://otheraccount.blob.core.windows.net/mycontainer/myblob?sastoken
-        :keyword bool overwrite: Whether the blob to be uploaded should overwrite the current data.
-            If True, upload_blob will overwrite the existing data. If set to False, the
-            operation will fail with ResourceExistsError.
-        :keyword bool include_source_blob_properties:
-            Indicates if properties from the source blob should be copied. Defaults to True.
-        :keyword tags:
-            Name-value pairs associated with the blob as tag. Tags are case-sensitive.
-            The tag set may contain at most 10 tags.  Tag keys must be between 1 and 128 characters,
-            and tag values must be between 0 and 256 characters.
-            Valid tag key and value characters include: lowercase and uppercase letters, digits (0-9),
-            space (` `), plus (+), minus (-), period (.), solidus (/), colon (:), equals (=), underscore (_)
-        :paramtype tags: dict(str, str)
-        :keyword bytearray source_content_md5:
-            Specify the md5 that is used to verify the integrity of the source bytes.
-        :keyword ~datetime.datetime source_if_modified_since:
-            A DateTime value. Azure expects the date value passed in to be UTC.
-            If timezone is included, any non-UTC datetimes will be converted to UTC.
-            If a date is passed in without timezone info, it is assumed to be UTC.
-            Specify this header to perform the operation only
-            if the source resource has been modified since the specified time.
-        :keyword ~datetime.datetime source_if_unmodified_since:
-            A DateTime value. Azure expects the date value passed in to be UTC.
-            If timezone is included, any non-UTC datetimes will be converted to UTC.
-            If a date is passed in without timezone info, it is assumed to be UTC.
-            Specify this header to perform the operation only if
-            the source resource has not been modified since the specified date/time.
-        :keyword str source_etag:
-            The source ETag value, or the wildcard character (*). Used to check if the resource has changed,
-            and act according to the condition specified by the `match_condition` parameter.
-        :keyword ~azure.core.MatchConditions source_match_condition:
-            The source match condition to use upon the etag.
-        :keyword ~datetime.datetime if_modified_since:
-            A DateTime value. Azure expects the date value passed in to be UTC.
-            If timezone is included, any non-UTC datetimes will be converted to UTC.
-            If a date is passed in without timezone info, it is assumed to be UTC.
-            Specify this header to perform the operation only
-            if the resource has been modified since the specified time.
-        :keyword ~datetime.datetime if_unmodified_since:
-            A DateTime value. Azure expects the date value passed in to be UTC.
-            If timezone is included, any non-UTC datetimes will be converted to UTC.
-            If a date is passed in without timezone info, it is assumed to be UTC.
-            Specify this header to perform the operation only if
-            the resource has not been modified since the specified date/time.
-        :keyword str etag:
-            The destination ETag value, or the wildcard character (*). Used to check if the resource has changed,
-            and act according to the condition specified by the `match_condition` parameter.
-        :keyword ~azure.core.MatchConditions match_condition:
-            The destination match condition to use upon the etag.
-        :keyword destination_lease:
-            The lease ID specified for this header must match the lease ID of the
-            destination blob. If the request does not include the lease ID or it is not
-            valid, the operation fails with status code 412 (Precondition Failed).
-        :paramtype destination_lease: ~azure.storage.blob.BlobLeaseClient or str
-        :keyword int timeout:
-            The timeout parameter is expressed in seconds.
-        :keyword ~azure.storage.blob.ContentSettings content_settings:
-            ContentSettings object used to set blob properties. Used to set content type, encoding,
-            language, disposition, md5, and cache control.
-        :keyword ~azure.storage.blob.CustomerProvidedEncryptionKey cpk:
-            Encrypts the data on the service-side with the given key.
-            Use of customer-provided keys must be done over HTTPS.
-            As the encryption key itself is provided in the request,
-            a secure connection must be established to transfer the key.
-        :keyword str encryption_scope:
-            A predefined encryption scope used to encrypt the data on the service. An encryption
-            scope can be created using the Management API and referenced here by name. If a default
-            encryption scope has been defined at the container, this value will override it if the
-            container-level scope is configured to allow overrides. Otherwise an error will be raised.
-        :keyword ~azure.storage.blob.StandardBlobTier standard_blob_tier:
-            A standard blob tier value to set the blob to. For this version of the library,
-            this is only applicable to block blobs on standard storage accounts.
-        """
-        options = self._upload_blob_from_url_options(
-            source_url=self._encode_source_url(source_url),
-            **kwargs)
-        try:
-            return self._client.block_blob.put_blob_from_url(**options)
-        except StorageErrorException as error:
-            process_storage_error(error)
-
     @distributed_trace
     def upload_blob(  # pylint: disable=too-many-locals
             self, data,  # type: Union[Iterable[AnyStr], IO[AnyStr]]
@@ -618,7 +478,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -778,7 +638,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -911,7 +771,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -1028,7 +888,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -1138,7 +998,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -1245,7 +1105,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -1319,7 +1179,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -1984,7 +1844,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -2032,7 +1892,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             This keyword argument was introduced in API version '2019-12-12'.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
         :keyword int timeout:
@@ -2481,7 +2341,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
         :type premium_page_blob_tier: ~azure.storage.blob.PremiumPageBlobTier
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -2512,12 +2372,10 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
     def _set_blob_tags_options(self, tags=None, **kwargs):
         # type: (Optional[Dict[str, str]], **Any) -> Dict[str, Any]
         tags = serialize_blob_tags(tags)
-        access_conditions = get_access_conditions(kwargs.pop('lease', None))
         mod_conditions = get_modify_conditions(kwargs)
 
         options = {
             'tags': tags,
-            'lease_access_conditions': access_conditions,
             'modified_access_conditions': mod_conditions,
             'cls': return_response_headers}
         options.update(kwargs)
@@ -2552,11 +2410,6 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             blob.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on destination blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
-        :keyword lease:
-            Required if the blob has an active lease. Value can be a BlobLeaseClient object
-            or the lease ID as a string.
-        :paramtype lease: ~azure.storage.blob.BlobLeaseClient or str
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
         :returns: Blob-updated property dict (Etag and last modified)
@@ -2570,13 +2423,11 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
 
     def _get_blob_tags_options(self, **kwargs):
         # type: (**Any) -> Dict[str, str]
-        access_conditions = get_access_conditions(kwargs.pop('lease', None))
         mod_conditions = get_modify_conditions(kwargs)
 
         options = {
             'version_id': kwargs.pop('version_id', None),
             'snapshot': self.snapshot,
-            'lease_access_conditions': access_conditions,
             'modified_access_conditions': mod_conditions,
             'timeout': kwargs.pop('timeout', None),
             'cls': return_headers_and_deserialized}
@@ -2595,11 +2446,6 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             value that, when present, specifies the version of the blob to add tags to.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on destination blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
-        :keyword lease:
-            Required if the blob has an active lease. Value can be a BlobLeaseClient object
-            or the lease ID as a string.
-        :paramtype lease: ~azure.storage.blob.BlobLeaseClient or str
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
         :returns: Key value pairs of blob tags.
@@ -2697,7 +2543,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -2846,7 +2692,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -2920,7 +2766,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -3050,7 +2896,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -3222,7 +3068,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The destination match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -3339,7 +3185,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -3469,7 +3315,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -3616,7 +3462,7 @@ class BlobClient(StorageAccountHostsMixin):  # pylint: disable=too-many-public-m
             The destination match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
