+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_blob_service_client.py
@@ -28,8 +28,7 @@ from ._generated import AzureBlobStorage, VERSION
 from ._generated.models import StorageErrorException, StorageServiceProperties, KeyInfo
 from ._container_client import ContainerClient
 from ._blob_client import BlobClient
-from ._models import ContainerPropertiesPaged
-from ._list_blobs_helper import FilteredBlobPaged
+from ._models import ContainerPropertiesPaged, FilteredBlobPaged
 from ._serialize import get_api_version
 from ._deserialize import service_stats_deserialize, service_properties_deserialize
 
@@ -48,7 +47,6 @@ if TYPE_CHECKING:
         CorsRule,
         RetentionPolicy,
         StaticWebsite,
-        FilteredBlob
     )
