+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/_configuration.py
@@ -40,7 +40,7 @@ class AzureBlobStorageConfiguration(Configuration):
         self.generate_client_request_id = True
 
         self.url = url
-        self.version = "2020-04-08"
+        self.version = "2020-02-10"
 
     def _configure(self, **kwargs):
         self.user_agent_policy = kwargs.get('user_agent_policy') or policies.UserAgentPolicy(**kwargs)
