+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/aio/_azure_blob_storage_async.py
@@ -56,7 +56,7 @@ class AzureBlobStorage(object):
         self._client = AsyncPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
-        self.api_version = '2020-04-08'
+        self.api_version = '2020-02-10'
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)
