+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/aio/operations_async/_blob_operations_async.py
@@ -448,7 +448,7 @@ class BlobOperations:
             return cls(response, None, response_headers)
     get_properties.metadata = {'url': '/{containerName}/{blob}'}
 
-    async def delete(self, snapshot=None, version_id=None, timeout=None, delete_snapshots=None, request_id=None, blob_delete_type=None, lease_access_conditions=None, modified_access_conditions=None, *, cls=None, **kwargs):
+    async def delete(self, snapshot=None, version_id=None, timeout=None, delete_snapshots=None, request_id=None, lease_access_conditions=None, modified_access_conditions=None, *, cls=None, **kwargs):
         """If the storage account's soft delete feature is disabled then, when a
         blob is deleted, it is permanently removed from the storage account. If
         the storage account's soft delete feature is enabled, then, when a blob
@@ -491,11 +491,6 @@ class BlobOperations:
          KB character limit that is recorded in the analytics logs when storage
          analytics logging is enabled.
         :type request_id: str
-        :param blob_delete_type: Optional.  Only possible value is
-         'permanent', which specifies to permanently delete a blob if blob soft
-         delete is enabled. Possible values include: 'Permanent'
-        :type blob_delete_type: str or
-         ~azure.storage.blob.models.BlobDeleteType
         :param lease_access_conditions: Additional parameters for the
          operation
         :type lease_access_conditions:
@@ -546,8 +541,6 @@ class BlobOperations:
             query_parameters['versionid'] = self._serialize.query("version_id", version_id, 'str')
         if timeout is not None:
             query_parameters['timeout'] = self._serialize.query("timeout", timeout, 'int', minimum=0)
-        if blob_delete_type is not None:
-            query_parameters['deletetype'] = self._serialize.query("blob_delete_type", blob_delete_type, 'BlobDeleteType')
 
         # Construct headers
         header_parameters = {}
@@ -1126,7 +1119,7 @@ class BlobOperations:
             header_parameters['x-ms-client-request-id'] = self._serialize.header("request_id", request_id, 'str')
         header_parameters['x-ms-expiry-option'] = self._serialize.header("expiry_options", expiry_options, 'str')
         if expires_on is not None:
-            header_parameters['x-ms-expiry-time'] = self._serialize.header("expires_on", expires_on, 'str')
+            header_parameters['x-ms-expiry-time'] = self._serialize.header("expires_on", expires_on, 'rfc-1123')
 
         # Construct and send request
         request = self._client.put(url, query_parameters, header_parameters)
@@ -2883,7 +2876,7 @@ class BlobOperations:
         return deserialized
     query.metadata = {'url': '/{containerName}/{blob}'}
 
-    async def get_tags(self, timeout=None, request_id=None, snapshot=None, version_id=None, modified_access_conditions=None, lease_access_conditions=None, *, cls=None, **kwargs):
+    async def get_tags(self, timeout=None, request_id=None, snapshot=None, version_id=None, modified_access_conditions=None, *, cls=None, **kwargs):
         """The Get Tags operation enables users to get the tags associated with a
         blob.
 
@@ -2910,10 +2903,6 @@ class BlobOperations:
          operation
         :type modified_access_conditions:
          ~azure.storage.blob.models.ModifiedAccessConditions
-        :param lease_access_conditions: Additional parameters for the
-         operation
-        :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
         :param callable cls: A custom type or function that will be passed the
          direct response
         :return: BlobTags or the result of cls(response)
@@ -2925,9 +2914,6 @@ class BlobOperations:
         if_tags = None
         if modified_access_conditions is not None:
             if_tags = modified_access_conditions.if_tags
-        lease_id = None
-        if lease_access_conditions is not None:
-            lease_id = lease_access_conditions.lease_id
 
         comp = "tags"
 
@@ -2956,8 +2942,6 @@ class BlobOperations:
             header_parameters['x-ms-client-request-id'] = self._serialize.header("request_id", request_id, 'str')
         if if_tags is not None:
             header_parameters['x-ms-if-tags'] = self._serialize.header("if_tags", if_tags, 'str')
-        if lease_id is not None:
-            header_parameters['x-ms-lease-id'] = self._serialize.header("lease_id", lease_id, 'str')
 
         # Construct and send request
         request = self._client.get(url, query_parameters, header_parameters)
@@ -2986,7 +2970,7 @@ class BlobOperations:
         return deserialized
     get_tags.metadata = {'url': '/{containerName}/{blob}'}
 
-    async def set_tags(self, timeout=None, version_id=None, transactional_content_md5=None, transactional_content_crc64=None, request_id=None, tags=None, modified_access_conditions=None, lease_access_conditions=None, *, cls=None, **kwargs):
+    async def set_tags(self, timeout=None, version_id=None, transactional_content_md5=None, transactional_content_crc64=None, request_id=None, tags=None, modified_access_conditions=None, *, cls=None, **kwargs):
         """The Set Tags operation enables users to set tags on a blob.
 
         :param timeout: The timeout parameter is expressed in seconds. For
@@ -3014,10 +2998,6 @@ class BlobOperations:
          operation
         :type modified_access_conditions:
          ~azure.storage.blob.models.ModifiedAccessConditions
-        :param lease_access_conditions: Additional parameters for the
-         operation
-        :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
         :param callable cls: A custom type or function that will be passed the
          direct response
         :return: None or the result of cls(response)
@@ -3029,9 +3009,6 @@ class BlobOperations:
         if_tags = None
         if modified_access_conditions is not None:
             if_tags = modified_access_conditions.if_tags
-        lease_id = None
-        if lease_access_conditions is not None:
-            lease_id = lease_access_conditions.lease_id
 
         comp = "tags"
 
@@ -3062,8 +3039,6 @@ class BlobOperations:
             header_parameters['x-ms-client-request-id'] = self._serialize.header("request_id", request_id, 'str')
         if if_tags is not None:
             header_parameters['x-ms-if-tags'] = self._serialize.header("if_tags", if_tags, 'str')
-        if lease_id is not None:
-            header_parameters['x-ms-lease-id'] = self._serialize.header("lease_id", lease_id, 'str')
 
         # Construct body
         if tags is not None:
