+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/aio/operations_async/_block_blob_operations_async.py
@@ -239,250 +239,6 @@ class BlockBlobOperations:
             return cls(response, None, response_headers)
     upload.metadata = {'url': '/{containerName}/{blob}'}
 
-    async def put_blob_from_url(self, content_length, copy_source, timeout=None, transactional_content_md5=None, metadata=None, tier=None, request_id=None, source_content_md5=None, blob_tags_string=None, copy_source_blob_properties=None, blob_http_headers=None, lease_access_conditions=None, cpk_info=None, cpk_scope_info=None, modified_access_conditions=None, source_modified_access_conditions=None, *, cls=None, **kwargs):
-        """The Put Blob from URL operation creates a new Block Blob where the
-        contents of the blob are read from a given URL.  This API is supported
-        beginning with the 2020-04-08 version. Partial updates are not
-        supported with Put Blob from URL; the content of an existing blob is
-        overwritten with the content of the new blob.  To perform partial
-        updates to a block blob’s contents using a source URL, use the Put
-        Block from URL API in conjunction with Put Block List.
-
-        :param content_length: The length of the request.
-        :type content_length: long
-        :param copy_source: Specifies the name of the source page blob
-         snapshot. This value is a URL of up to 2 KB in length that specifies a
-         page blob snapshot. The value should be URL-encoded as it would appear
-         in a request URI. The source blob must either be public or must be
-         authenticated via a shared access signature.
-        :type copy_source: str
-        :param timeout: The timeout parameter is expressed in seconds. For
-         more information, see <a
-         href="https://docs.microsoft.com/en-us/rest/api/storageservices/fileservices/setting-timeouts-for-blob-service-operations">Setting
-         Timeouts for Blob Service Operations.</a>
-        :type timeout: int
-        :param transactional_content_md5: Specify the transactional md5 for
-         the body, to be validated by the service.
-        :type transactional_content_md5: bytearray
-        :param metadata: Optional. Specifies a user-defined name-value pair
-         associated with the blob. If no name-value pairs are specified, the
-         operation will copy the metadata from the source blob or file to the
-         destination blob. If one or more name-value pairs are specified, the
-         destination blob is created with the specified metadata, and metadata
-         is not copied from the source blob or file. Note that beginning with
-         version 2009-09-19, metadata names must adhere to the naming rules for
-         C# identifiers. See Naming and Referencing Containers, Blobs, and
-         Metadata for more information.
-        :type metadata: str
-        :param tier: Optional. Indicates the tier to be set on the blob.
-         Possible values include: 'P4', 'P6', 'P10', 'P15', 'P20', 'P30',
-         'P40', 'P50', 'P60', 'P70', 'P80', 'Hot', 'Cool', 'Archive'
-        :type tier: str or ~azure.storage.blob.models.AccessTierOptional
-        :param request_id: Provides a client-generated, opaque value with a 1
-         KB character limit that is recorded in the analytics logs when storage
-         analytics logging is enabled.
-        :type request_id: str
-        :param source_content_md5: Specify the md5 calculated for the range of
-         bytes that must be read from the copy source.
-        :type source_content_md5: bytearray
-        :param blob_tags_string: Optional.  Used to set blob tags in various
-         blob operations.
-        :type blob_tags_string: str
-        :param copy_source_blob_properties: Optional, default is true.
-         Indicates if properties from the source blob should be copied.
-        :type copy_source_blob_properties: bool
-        :param blob_http_headers: Additional parameters for the operation
-        :type blob_http_headers: ~azure.storage.blob.models.BlobHTTPHeaders
-        :param lease_access_conditions: Additional parameters for the
-         operation
-        :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
-        :param cpk_info: Additional parameters for the operation
-        :type cpk_info: ~azure.storage.blob.models.CpkInfo
-        :param cpk_scope_info: Additional parameters for the operation
-        :type cpk_scope_info: ~azure.storage.blob.models.CpkScopeInfo
-        :param modified_access_conditions: Additional parameters for the
-         operation
-        :type modified_access_conditions:
-         ~azure.storage.blob.models.ModifiedAccessConditions
-        :param source_modified_access_conditions: Additional parameters for
-         the operation
-        :type source_modified_access_conditions:
-         ~azure.storage.blob.models.SourceModifiedAccessConditions
-        :param callable cls: A custom type or function that will be passed the
-         direct response
-        :return: None or the result of cls(response)
-        :rtype: None
-        :raises:
-         :class:`StorageErrorException<azure.storage.blob.models.StorageErrorException>`
-        """
-        error_map = kwargs.pop('error_map', None)
-        blob_content_type = None
-        if blob_http_headers is not None:
-            blob_content_type = blob_http_headers.blob_content_type
-        blob_content_encoding = None
-        if blob_http_headers is not None:
-            blob_content_encoding = blob_http_headers.blob_content_encoding
-        blob_content_language = None
-        if blob_http_headers is not None:
-            blob_content_language = blob_http_headers.blob_content_language
-        blob_content_md5 = None
-        if blob_http_headers is not None:
-            blob_content_md5 = blob_http_headers.blob_content_md5
-        blob_cache_control = None
-        if blob_http_headers is not None:
-            blob_cache_control = blob_http_headers.blob_cache_control
-        blob_content_disposition = None
-        if blob_http_headers is not None:
-            blob_content_disposition = blob_http_headers.blob_content_disposition
-        lease_id = None
-        if lease_access_conditions is not None:
-            lease_id = lease_access_conditions.lease_id
-        encryption_key = None
-        if cpk_info is not None:
-            encryption_key = cpk_info.encryption_key
-        encryption_key_sha256 = None
-        if cpk_info is not None:
-            encryption_key_sha256 = cpk_info.encryption_key_sha256
-        encryption_algorithm = None
-        if cpk_info is not None:
-            encryption_algorithm = cpk_info.encryption_algorithm
-        encryption_scope = None
-        if cpk_scope_info is not None:
-            encryption_scope = cpk_scope_info.encryption_scope
-        if_modified_since = None
-        if modified_access_conditions is not None:
-            if_modified_since = modified_access_conditions.if_modified_since
-        if_unmodified_since = None
-        if modified_access_conditions is not None:
-            if_unmodified_since = modified_access_conditions.if_unmodified_since
-        if_match = None
-        if modified_access_conditions is not None:
-            if_match = modified_access_conditions.if_match
-        if_none_match = None
-        if modified_access_conditions is not None:
-            if_none_match = modified_access_conditions.if_none_match
-        if_tags = None
-        if modified_access_conditions is not None:
-            if_tags = modified_access_conditions.if_tags
-        source_if_modified_since = None
-        if source_modified_access_conditions is not None:
-            source_if_modified_since = source_modified_access_conditions.source_if_modified_since
-        source_if_unmodified_since = None
-        if source_modified_access_conditions is not None:
-            source_if_unmodified_since = source_modified_access_conditions.source_if_unmodified_since
-        source_if_match = None
-        if source_modified_access_conditions is not None:
-            source_if_match = source_modified_access_conditions.source_if_match
-        source_if_none_match = None
-        if source_modified_access_conditions is not None:
-            source_if_none_match = source_modified_access_conditions.source_if_none_match
-        source_if_tags = None
-        if source_modified_access_conditions is not None:
-            source_if_tags = source_modified_access_conditions.source_if_tags
-
-        # Construct URL
-        url = self.put_blob_from_url.metadata['url']
-        path_format_arguments = {
-            'url': self._serialize.url("self._config.url", self._config.url, 'str', skip_quote=True)
-        }
-        url = self._client.format_url(url, **path_format_arguments)
-
-        # Construct parameters
-        query_parameters = {}
-        if timeout is not None:
-            query_parameters['timeout'] = self._serialize.query("timeout", timeout, 'int', minimum=0)
-
-        # Construct headers
-        header_parameters = {}
-        if transactional_content_md5 is not None:
-            header_parameters['Content-MD5'] = self._serialize.header("transactional_content_md5", transactional_content_md5, 'bytearray')
-        header_parameters['Content-Length'] = self._serialize.header("content_length", content_length, 'long')
-        if metadata is not None:
-            header_parameters['x-ms-meta'] = self._serialize.header("metadata", metadata, 'str')
-        if tier is not None:
-            header_parameters['x-ms-access-tier'] = self._serialize.header("tier", tier, 'str')
-        header_parameters['x-ms-version'] = self._serialize.header("self._config.version", self._config.version, 'str')
-        if request_id is not None:
-            header_parameters['x-ms-client-request-id'] = self._serialize.header("request_id", request_id, 'str')
-        if source_content_md5 is not None:
-            header_parameters['x-ms-source-content-md5'] = self._serialize.header("source_content_md5", source_content_md5, 'bytearray')
-        if blob_tags_string is not None:
-            header_parameters['x-ms-tags'] = self._serialize.header("blob_tags_string", blob_tags_string, 'str')
-        header_parameters['x-ms-copy-source'] = self._serialize.header("copy_source", copy_source, 'str')
-        if copy_source_blob_properties is not None:
-            header_parameters['x-ms-copy-source-blob-properties'] = self._serialize.header("copy_source_blob_properties", copy_source_blob_properties, 'bool')
-        header_parameters['x-ms-blob-type'] = self._serialize.header("self.x_ms_blob_type", self.x_ms_blob_type, 'str')
-        if blob_content_type is not None:
-            header_parameters['x-ms-blob-content-type'] = self._serialize.header("blob_content_type", blob_content_type, 'str')
-        if blob_content_encoding is not None:
-            header_parameters['x-ms-blob-content-encoding'] = self._serialize.header("blob_content_encoding", blob_content_encoding, 'str')
-        if blob_content_language is not None:
-            header_parameters['x-ms-blob-content-language'] = self._serialize.header("blob_content_language", blob_content_language, 'str')
-        if blob_content_md5 is not None:
-            header_parameters['x-ms-blob-content-md5'] = self._serialize.header("blob_content_md5", blob_content_md5, 'bytearray')
-        if blob_cache_control is not None:
-            header_parameters['x-ms-blob-cache-control'] = self._serialize.header("blob_cache_control", blob_cache_control, 'str')
-        if blob_content_disposition is not None:
-            header_parameters['x-ms-blob-content-disposition'] = self._serialize.header("blob_content_disposition", blob_content_disposition, 'str')
-        if lease_id is not None:
-            header_parameters['x-ms-lease-id'] = self._serialize.header("lease_id", lease_id, 'str')
-        if encryption_key is not None:
-            header_parameters['x-ms-encryption-key'] = self._serialize.header("encryption_key", encryption_key, 'str')
-        if encryption_key_sha256 is not None:
-            header_parameters['x-ms-encryption-key-sha256'] = self._serialize.header("encryption_key_sha256", encryption_key_sha256, 'str')
-        if encryption_algorithm is not None:
-            header_parameters['x-ms-encryption-algorithm'] = self._serialize.header("encryption_algorithm", encryption_algorithm, 'EncryptionAlgorithmType')
-        if encryption_scope is not None:
-            header_parameters['x-ms-encryption-scope'] = self._serialize.header("encryption_scope", encryption_scope, 'str')
-        if if_modified_since is not None:
-            header_parameters['If-Modified-Since'] = self._serialize.header("if_modified_since", if_modified_since, 'rfc-1123')
-        if if_unmodified_since is not None:
-            header_parameters['If-Unmodified-Since'] = self._serialize.header("if_unmodified_since", if_unmodified_since, 'rfc-1123')
-        if if_match is not None:
-            header_parameters['If-Match'] = self._serialize.header("if_match", if_match, 'str')
-        if if_none_match is not None:
-            header_parameters['If-None-Match'] = self._serialize.header("if_none_match", if_none_match, 'str')
-        if if_tags is not None:
-            header_parameters['x-ms-if-tags'] = self._serialize.header("if_tags", if_tags, 'str')
-        if source_if_modified_since is not None:
-            header_parameters['x-ms-source-if-modified-since'] = self._serialize.header("source_if_modified_since", source_if_modified_since, 'rfc-1123')
-        if source_if_unmodified_since is not None:
-            header_parameters['x-ms-source-if-unmodified-since'] = self._serialize.header("source_if_unmodified_since", source_if_unmodified_since, 'rfc-1123')
-        if source_if_match is not None:
-            header_parameters['x-ms-source-if-match'] = self._serialize.header("source_if_match", source_if_match, 'str')
-        if source_if_none_match is not None:
-            header_parameters['x-ms-source-if-none-match'] = self._serialize.header("source_if_none_match", source_if_none_match, 'str')
-        if source_if_tags is not None:
-            header_parameters['x-ms-source-if-tags'] = self._serialize.header("source_if_tags", source_if_tags, 'str')
-
-        # Construct and send request
-        request = self._client.put(url, query_parameters, header_parameters)
-        pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
-        response = pipeline_response.http_response
-
-        if response.status_code not in [201]:
-            map_error(status_code=response.status_code, response=response, error_map=error_map)
-            raise models.StorageErrorException(response, self._deserialize)
-
-        if cls:
-            response_headers = {
-                'ETag': self._deserialize('str', response.headers.get('ETag')),
-                'Last-Modified': self._deserialize('rfc-1123', response.headers.get('Last-Modified')),
-                'Content-MD5': self._deserialize('bytearray', response.headers.get('Content-MD5')),
-                'x-ms-client-request-id': self._deserialize('str', response.headers.get('x-ms-client-request-id')),
-                'x-ms-request-id': self._deserialize('str', response.headers.get('x-ms-request-id')),
-                'x-ms-version': self._deserialize('str', response.headers.get('x-ms-version')),
-                'x-ms-version-id': self._deserialize('str', response.headers.get('x-ms-version-id')),
-                'Date': self._deserialize('rfc-1123', response.headers.get('Date')),
-                'x-ms-request-server-encrypted': self._deserialize('bool', response.headers.get('x-ms-request-server-encrypted')),
-                'x-ms-encryption-key-sha256': self._deserialize('str', response.headers.get('x-ms-encryption-key-sha256')),
-                'x-ms-encryption-scope': self._deserialize('str', response.headers.get('x-ms-encryption-scope')),
-                'x-ms-error-code': self._deserialize('str', response.headers.get('x-ms-error-code')),
-            }
-            return cls(response, None, response_headers)
-    put_blob_from_url.metadata = {'url': '/{containerName}/{blob}'}
-
     async def stage_block(self, block_id, content_length, body, transactional_content_md5=None, transactional_content_crc64=None, timeout=None, request_id=None, lease_access_conditions=None, cpk_info=None, cpk_scope_info=None, *, cls=None, **kwargs):
         """The Stage Block operation creates a new block to be committed as part
         of a blob.
