+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/models/__init__.py
@@ -123,7 +123,6 @@ from ._azure_blob_storage_enums import (
     AccessTierRequired,
     AccountKind,
     ArchiveStatus,
-    BlobDeleteType,
     BlobExpiryOptions,
     BlobType,
     BlockListType,
@@ -216,7 +215,6 @@ __all__ = [
     'AccessTierRequired',
     'AccessTierOptional',
     'PremiumPageBlobAccessTier',
-    'BlobDeleteType',
     'BlobExpiryOptions',
     'BlockListType',
     'DeleteSnapshotsOptionType',
