+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/models/_azure_blob_storage_enums.py
@@ -130,7 +130,6 @@ class StorageErrorCode(str, Enum):
     unsupported_http_verb = "UnsupportedHttpVerb"
     append_position_condition_not_met = "AppendPositionConditionNotMet"
     blob_already_exists = "BlobAlreadyExists"
-    blob_immutable_due_to_policy = "BlobImmutableDueToPolicy"
     blob_not_found = "BlobNotFound"
     blob_overwritten = "BlobOverwritten"
     blob_tier_inadequate_for_content_length = "BlobTierInadequateForContentLength"
@@ -263,11 +262,6 @@ class PremiumPageBlobAccessTier(str, Enum):
     p60 = "P60"
     p70 = "P70"
     p80 = "P80"
-
-
-class BlobDeleteType(str, Enum):
-
-    permanent = "Permanent"
 
 
 class BlobExpiryOptions(str, Enum):
