+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/models/_models.py
@@ -1084,19 +1084,20 @@ class FilterBlobItem(Model):
     :type name: str
     :param container_name: Required.
     :type container_name: str
-    :param tags:
-    :type tags: ~azure.storage.blob.models.BlobTags
+    :param tag_value: Required.
+    :type tag_value: str
     """
 
     _validation = {
         'name': {'required': True},
         'container_name': {'required': True},
+        'tag_value': {'required': True},
     }
 
     _attribute_map = {
         'name': {'key': 'Name', 'type': 'str', 'xml': {'name': 'Name'}},
         'container_name': {'key': 'ContainerName', 'type': 'str', 'xml': {'name': 'ContainerName'}},
-        'tags': {'key': 'Tags', 'type': 'BlobTags', 'xml': {'name': 'Tags'}},
+        'tag_value': {'key': 'TagValue', 'type': 'str', 'xml': {'name': 'TagValue'}},
     }
     _xml_map = {
         'name': 'Blob'
@@ -1106,7 +1107,7 @@ class FilterBlobItem(Model):
         super(FilterBlobItem, self).__init__(**kwargs)
         self.name = kwargs.get('name', None)
         self.container_name = kwargs.get('container_name', None)
-        self.tags = kwargs.get('tags', None)
+        self.tag_value = kwargs.get('tag_value', None)
 
 
 class FilterBlobSegment(Model):
@@ -1696,9 +1697,6 @@ class RetentionPolicy(Model):
      soft-deleted data should be retained. All data older than this value will
      be deleted
     :type days: int
-    :param allow_permanent_delete: Indicates whether permanent delete is
-     allowed on this storage account.
-    :type allow_permanent_delete: bool
     """
 
     _validation = {
@@ -1709,7 +1707,6 @@ class RetentionPolicy(Model):
     _attribute_map = {
         'enabled': {'key': 'Enabled', 'type': 'bool', 'xml': {'name': 'Enabled'}},
         'days': {'key': 'Days', 'type': 'int', 'xml': {'name': 'Days'}},
-        'allow_permanent_delete': {'key': 'AllowPermanentDelete', 'type': 'bool', 'xml': {'name': 'AllowPermanentDelete'}},
     }
     _xml_map = {
     }
@@ -1718,7 +1715,6 @@ class RetentionPolicy(Model):
         super(RetentionPolicy, self).__init__(**kwargs)
         self.enabled = kwargs.get('enabled', None)
         self.days = kwargs.get('days', None)
-        self.allow_permanent_delete = kwargs.get('allow_permanent_delete', None)
 
 
 class SequenceNumberAccessConditions(Model):
