+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/models/_models_py3.py
@@ -1084,29 +1084,30 @@ class FilterBlobItem(Model):
     :type name: str
     :param container_name: Required.
     :type container_name: str
-    :param tags:
-    :type tags: ~azure.storage.blob.models.BlobTags
+    :param tag_value: Required.
+    :type tag_value: str
     """
 
     _validation = {
         'name': {'required': True},
         'container_name': {'required': True},
+        'tag_value': {'required': True},
     }
 
     _attribute_map = {
         'name': {'key': 'Name', 'type': 'str', 'xml': {'name': 'Name'}},
         'container_name': {'key': 'ContainerName', 'type': 'str', 'xml': {'name': 'ContainerName'}},
-        'tags': {'key': 'Tags', 'type': 'BlobTags', 'xml': {'name': 'Tags'}},
+        'tag_value': {'key': 'TagValue', 'type': 'str', 'xml': {'name': 'TagValue'}},
     }
     _xml_map = {
         'name': 'Blob'
     }
 
-    def __init__(self, *, name: str, container_name: str, tags=None, **kwargs) -> None:
+    def __init__(self, *, name: str, container_name: str, tag_value: str, **kwargs) -> None:
         super(FilterBlobItem, self).__init__(**kwargs)
         self.name = name
         self.container_name = container_name
-        self.tags = tags
+        self.tag_value = tag_value
 
 
 class FilterBlobSegment(Model):
@@ -1696,9 +1697,6 @@ class RetentionPolicy(Model):
      soft-deleted data should be retained. All data older than this value will
      be deleted
     :type days: int
-    :param allow_permanent_delete: Indicates whether permanent delete is
-     allowed on this storage account.
-    :type allow_permanent_delete: bool
     """
 
     _validation = {
@@ -1709,16 +1707,14 @@ class RetentionPolicy(Model):
     _attribute_map = {
         'enabled': {'key': 'Enabled', 'type': 'bool', 'xml': {'name': 'Enabled'}},
         'days': {'key': 'Days', 'type': 'int', 'xml': {'name': 'Days'}},
-        'allow_permanent_delete': {'key': 'AllowPermanentDelete', 'type': 'bool', 'xml': {'name': 'AllowPermanentDelete'}},
     }
     _xml_map = {
     }
 
-    def __init__(self, *, enabled: bool, days: int=None, allow_permanent_delete: bool=None, **kwargs) -> None:
+    def __init__(self, *, enabled: bool, days: int=None, **kwargs) -> None:
         super(RetentionPolicy, self).__init__(**kwargs)
         self.enabled = enabled
         self.days = days
-        self.allow_permanent_delete = allow_permanent_delete
 
 
 class SequenceNumberAccessConditions(Model):
