+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/operations/_blob_operations.py
@@ -447,7 +447,7 @@ class BlobOperations(object):
             return cls(response, None, response_headers)
     get_properties.metadata = {'url': '/{containerName}/{blob}'}
 
-    def delete(self, snapshot=None, version_id=None, timeout=None, delete_snapshots=None, request_id=None, blob_delete_type=None, lease_access_conditions=None, modified_access_conditions=None, cls=None, **kwargs):
+    def delete(self, snapshot=None, version_id=None, timeout=None, delete_snapshots=None, request_id=None, lease_access_conditions=None, modified_access_conditions=None, cls=None, **kwargs):
         """If the storage account's soft delete feature is disabled then, when a
         blob is deleted, it is permanently removed from the storage account. If
         the storage account's soft delete feature is enabled, then, when a blob
@@ -490,11 +490,6 @@ class BlobOperations(object):
          KB character limit that is recorded in the analytics logs when storage
          analytics logging is enabled.
         :type request_id: str
-        :param blob_delete_type: Optional.  Only possible value is
-         'permanent', which specifies to permanently delete a blob if blob soft
-         delete is enabled. Possible values include: 'Permanent'
-        :type blob_delete_type: str or
-         ~azure.storage.blob.models.BlobDeleteType
         :param lease_access_conditions: Additional parameters for the
          operation
         :type lease_access_conditions:
@@ -545,8 +540,6 @@ class BlobOperations(object):
             query_parameters['versionid'] = self._serialize.query("version_id", version_id, 'str')
         if timeout is not None:
             query_parameters['timeout'] = self._serialize.query("timeout", timeout, 'int', minimum=0)
-        if blob_delete_type is not None:
-            query_parameters['deletetype'] = self._serialize.query("blob_delete_type", blob_delete_type, 'BlobDeleteType')
 
         # Construct headers
         header_parameters = {}
@@ -1125,7 +1118,7 @@ class BlobOperations(object):
             header_parameters['x-ms-client-request-id'] = self._serialize.header("request_id", request_id, 'str')
         header_parameters['x-ms-expiry-option'] = self._serialize.header("expiry_options", expiry_options, 'str')
         if expires_on is not None:
-            header_parameters['x-ms-expiry-time'] = self._serialize.header("expires_on", expires_on, 'str')
+            header_parameters['x-ms-expiry-time'] = self._serialize.header("expires_on", expires_on, 'rfc-1123')
 
         # Construct and send request
         request = self._client.put(url, query_parameters, header_parameters)
@@ -2881,7 +2874,7 @@ class BlobOperations(object):
         return deserialized
     query.metadata = {'url': '/{containerName}/{blob}'}
 
-    def get_tags(self, timeout=None, request_id=None, snapshot=None, version_id=None, modified_access_conditions=None, lease_access_conditions=None, cls=None, **kwargs):
+    def get_tags(self, timeout=None, request_id=None, snapshot=None, version_id=None, modified_access_conditions=None, cls=None, **kwargs):
         """The Get Tags operation enables users to get the tags associated with a
         blob.
 
@@ -2908,10 +2901,6 @@ class BlobOperations(object):
          operation
         :type modified_access_conditions:
          ~azure.storage.blob.models.ModifiedAccessConditions
-        :param lease_access_conditions: Additional parameters for the
-         operation
-        :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
         :param callable cls: A custom type or function that will be passed the
          direct response
         :return: BlobTags or the result of cls(response)
@@ -2923,9 +2912,6 @@ class BlobOperations(object):
         if_tags = None
         if modified_access_conditions is not None:
             if_tags = modified_access_conditions.if_tags
-        lease_id = None
-        if lease_access_conditions is not None:
-            lease_id = lease_access_conditions.lease_id
 
         comp = "tags"
 
@@ -2954,8 +2940,6 @@ class BlobOperations(object):
             header_parameters['x-ms-client-request-id'] = self._serialize.header("request_id", request_id, 'str')
         if if_tags is not None:
             header_parameters['x-ms-if-tags'] = self._serialize.header("if_tags", if_tags, 'str')
-        if lease_id is not None:
-            header_parameters['x-ms-lease-id'] = self._serialize.header("lease_id", lease_id, 'str')
 
         # Construct and send request
         request = self._client.get(url, query_parameters, header_parameters)
@@ -2984,7 +2968,7 @@ class BlobOperations(object):
         return deserialized
     get_tags.metadata = {'url': '/{containerName}/{blob}'}
 
-    def set_tags(self, timeout=None, version_id=None, transactional_content_md5=None, transactional_content_crc64=None, request_id=None, tags=None, modified_access_conditions=None, lease_access_conditions=None, cls=None, **kwargs):
+    def set_tags(self, timeout=None, version_id=None, transactional_content_md5=None, transactional_content_crc64=None, request_id=None, tags=None, modified_access_conditions=None, cls=None, **kwargs):
         """The Set Tags operation enables users to set tags on a blob.
 
         :param timeout: The timeout parameter is expressed in seconds. For
@@ -3012,10 +2996,6 @@ class BlobOperations(object):
          operation
         :type modified_access_conditions:
          ~azure.storage.blob.models.ModifiedAccessConditions
-        :param lease_access_conditions: Additional parameters for the
-         operation
-        :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
         :param callable cls: A custom type or function that will be passed the
          direct response
         :return: None or the result of cls(response)
@@ -3027,9 +3007,6 @@ class BlobOperations(object):
         if_tags = None
         if modified_access_conditions is not None:
             if_tags = modified_access_conditions.if_tags
-        lease_id = None
-        if lease_access_conditions is not None:
-            lease_id = lease_access_conditions.lease_id
 
         comp = "tags"
 
@@ -3060,8 +3037,6 @@ class BlobOperations(object):
             header_parameters['x-ms-client-request-id'] = self._serialize.header("request_id", request_id, 'str')
         if if_tags is not None:
             header_parameters['x-ms-if-tags'] = self._serialize.header("if_tags", if_tags, 'str')
-        if lease_id is not None:
-            header_parameters['x-ms-lease-id'] = self._serialize.header("lease_id", lease_id, 'str')
 
         # Construct body
         if tags is not None:
