+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_generated/version.py
@@ -9,5 +9,5 @@
 # regenerated.
 # --------------------------------------------------------------------------
 
-VERSION = "2020-04-08"
+VERSION = "2020-02-10"
