+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_lease.py
@@ -98,7 +98,7 @@ class BlobLeaseClient(object):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -151,7 +151,7 @@ class BlobLeaseClient(object):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -201,7 +201,7 @@ class BlobLeaseClient(object):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -250,7 +250,7 @@ class BlobLeaseClient(object):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -309,7 +309,7 @@ class BlobLeaseClient(object):
             the resource has not been modified since the specified date/time.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
