+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_serialize.py
@@ -36,7 +36,6 @@ _SUPPORTED_API_VERSIONS = [
     '2019-10-10',
     '2019-12-12',
     '2020-02-10',
-    '2020-04-08'
 ]
