+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/_version.py
@@ -4,4 +4,4 @@
 # license information.
 # --------------------------------------------------------------------------
 
-VERSION = "12.7.0b1"
+VERSION = "12.6.0"
