+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/aio/_container_client_async.py
@@ -659,7 +659,7 @@ class ContainerClient(AsyncStorageAccountHostsMixin, ContainerClientBase):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -782,7 +782,7 @@ class ContainerClient(AsyncStorageAccountHostsMixin, ContainerClientBase):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -847,7 +847,7 @@ class ContainerClient(AsyncStorageAccountHostsMixin, ContainerClientBase):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -936,7 +936,7 @@ class ContainerClient(AsyncStorageAccountHostsMixin, ContainerClientBase):
             the resource has not been modified since the specified date/time.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -1014,7 +1014,7 @@ class ContainerClient(AsyncStorageAccountHostsMixin, ContainerClientBase):
             Indicates the priority with which to rehydrate an archived blob
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
