+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/aio/_lease_async.py
@@ -94,7 +94,7 @@ class BlobLeaseClient(LeaseClientBase):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -147,7 +147,7 @@ class BlobLeaseClient(LeaseClientBase):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -197,7 +197,7 @@ class BlobLeaseClient(LeaseClientBase):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -246,7 +246,7 @@ class BlobLeaseClient(LeaseClientBase):
             The match condition to use upon the etag.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
 
@@ -305,7 +305,7 @@ class BlobLeaseClient(LeaseClientBase):
             the resource has not been modified since the specified date/time.
         :keyword str if_tags_match_condition:
             Specify a SQL where clause on blob tags to operate only on blob with a matching value.
-            eg. ``\"\\\"tagname\\\"='my tag'\"``
+            eg. "\"tagname\"='my tag'"
 
             .. versionadded:: 12.4.0
