+++ src/storage-preview/azext_storage_preview/vendored_sdks/blob/aio/_models.py
@@ -7,7 +7,6 @@
 # pylint: disable=super-init-not-called, too-many-lines
 
 from azure.core.async_paging import AsyncPageIterator
-from .._deserialize import parse_tags
 
 from .._models import ContainerProperties, FilteredBlob
 from .._shared.response_handlers import return_context_and_deserialized, process_storage_error
@@ -137,7 +136,6 @@ class FilteredBlobPaged(AsyncPageIterator):
     @staticmethod
     def _build_item(item):
         if isinstance(item, FilterBlobItem):
-            tags = parse_tags(item.tags)
-            blob = FilteredBlob(name=item.name, container_name=item.container_name, tags=tags)
+            blob = FilteredBlob(name=item.name, container_name=item.container_name, tag_value=item.tag_value)  # pylint: disable=protected-access
             return blob
         return item
